Updating type for AT num_instr to 32 bit int

+++ R2U2_C/src/AT/at_globals.h
@@ -12,7 +12,7 @@ typedef at_instruction_t instructions_t[N_AT];
 /* similar to TL_observers? */
 extern signals_vector_t signals_vector;
 extern instructions_t at_instructions;
-extern uint8_t num_instr;
+extern uint32_t num_instr;
 
 /* For no file handling option */
 extern char *at_bin;
